Fix collision bug in slopes in deli
This fixes a collision bug where you get stuck on the slopes near the
bases in deli (the last slope inside the house).  The bug happened
because you collide with the block below the slope.  I fixed it by
removing two blocks below each slope.  This fix is completely invisible
since the blocks I removed were not visible.